Refactoring the Stress-stiffening classes
